Move delegate signing into the utils package
so workers don’t have to import the whole sdk
