+++ packages/utils/package.json
@@ -19,6 +19,8 @@
     },
     "dependencies": {
         "@bufbuild/protobuf": "^2.9.0",
+        "@ethereumjs/util": "^10.0.0",
+        "@noble/hashes": "^1.8.0",
         "@towns-protocol/proto": "workspace:^",
         "debug": "^4.3.4",
         "ethereum-cryptography": "^3.2.0"
