+++ packages/utils/src/index.ts
@@ -3,6 +3,7 @@
  **************************************************************************/
 export * from './binary'
 export * from './check'
+export * from './delegate'
 export * from './dlog'
 export * from './envUtils'
 export * from './utils'
